improve batch adapters
